requirements.txt
fastapi
meraki
loguru
pydantic
python-dotenv
requests